+++ packages/camera/camera_android/lib/src/android_camera.dart
@@ -381,6 +381,10 @@ class AndroidCamera extends CameraPlatform {
   }
 
   @override
+  Future<void> setImageQuality(int cameraId, int quality) =>
+      _hostApi.setImageQuality(quality);
+
+  @override
   Widget buildPreview(int cameraId) {
     return Texture(textureId: cameraId);
   }