+++ packages/camera/camera_android_camerax/pigeons/camerax_library.dart
@@ -598,7 +598,7 @@ enum CameraXFlashMode {
   ),
 )
 abstract class ImageCapture extends UseCase {
-  ImageCapture(int? targetRotation, CameraXFlashMode? flashMode);
+  ImageCapture(int? targetRotation, CameraXFlashMode? flashMode, int? jpegQuality);
 
   late final ResolutionSelector? resolutionSelector;