+++ packages/camera/camera_platform_interface/test/camera_platform_interface_test.dart
@@ -521,6 +521,20 @@ void main() {
     );
 
     test(
+      'Default implementation of setImageQuality() should throw unimplemented error',
+      () {
+        // Arrange
+        final cameraPlatform = ExtendsCameraPlatform();
+
+        // Act & Assert
+        expect(
+          () => cameraPlatform.setImageQuality(1, 50),
+          throwsUnimplementedError,
+        );
+      },
+    );
+
+    test(
       'Default implementation of supportsImageStreaming() should return false',
       () {
         // Arrange