+++ packages/camera/camera_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.10.11
+
+* Adds `setImageQuality` for controlling JPEG compression quality.
+
 ## 0.10.10+16
 
 * Updates build files from Groovy to Kotlin.